moved from modMcf to modAion: IChainCfg, IRewardsCalculator, AbstractPowBlockstore, IBlockStorePow, IPowListener, removed unused IMiner reference, deleted empty interface EvtData
