+++ modAion/src/org/aion/zero/db/AbstractPowBlockstore.java
@@ -20,41 +20,33 @@
  *
  * Contributors:
  *     Aion foundation.
-
+ *
  ******************************************************************************/
-package org.aion.mcf.evt;
+package org.aion.zero.db;
 
-import java.util.List;
-
-import org.aion.base.type.IBlock;
-import org.aion.base.type.ITransaction;
-import org.aion.base.type.ITxExecSummary;
-
-import org.aion.mcf.types.AbstractBlockSummary;
-import org.aion.mcf.types.AbstractTxReceipt;
+import org.aion.mcf.types.AbstractBlock;
+import org.aion.mcf.types.AbstractBlockHeader;
 
 /**
- * POW listener interface.
+ * Abstract POW blockstore.
  *
  * @param <BLK>
- * @param <TX>
- * @param <TXR>
- * @param <BS>
+ * @param <BH>
  */
-public interface IPowListener<BLK extends IBlock<?, ?>, TX extends ITransaction, TXR extends AbstractTxReceipt<?>, BS extends AbstractBlockSummary<?, ?, ?, ?>>
-        extends IListenerBase<BLK, TX, TXR, BS> {
-    void onBlock(BS blockSummary);
+public abstract class AbstractPowBlockstore<BLK extends AbstractBlock<?, ?>, BH extends AbstractBlockHeader>
+        implements IBlockStorePow<BLK, BH> {
 
-    void onPeerDisconnect(String host, long port);
-
-    void onPendingTransactionsReceived(List<TX> transactions);
-
-    void onSyncDone();
-
-    void onNoConnections();
-
-    void onVMTraceCreated(String transactionHash, String trace);
-
-    void onTransactionExecuted(ITxExecSummary summary);
+    @Override
+    public byte[] getBlockHashByNumber(long blockNumber, byte[] branchBlockHash) {
+        BLK branchBlock = getBlockByHash(branchBlockHash);
+        if (branchBlock.getNumber() < blockNumber) {
+            throw new IllegalArgumentException(
+                    "Requested block number > branch hash number: " + blockNumber + " < " + branchBlock.getNumber());
+        }
+        while (branchBlock.getNumber() > blockNumber) {
+            branchBlock = getBlockByHash(branchBlock.getParentHash());
+        }
+        return branchBlock.getHash();
+    }
 
 }
