+++ modAion/src/org/aion/zero/db/IBlockStorePow.java
@@ -20,13 +20,28 @@
  *
  * Contributors:
  *     Aion foundation.
-
  ******************************************************************************/
-package org.aion.mcf.evt;
+package org.aion.zero.db;
+
+import java.math.BigInteger;
+
+import org.aion.base.type.IBlock;
+import org.aion.mcf.db.IBlockStoreBase;
+import org.aion.mcf.types.AbstractBlockHeader;
 
 /**
- * @author jin
+ * POW BLockstore interface.
+ *
+ * @param <BLK>
+ * @param <BH>
  */
-public interface EvtData {
+public interface IBlockStorePow<BLK extends IBlock<?, ?>, BH extends AbstractBlockHeader>
+        extends IBlockStoreBase<BLK, BH> {
+
+    BigInteger getTotalDifficultyForHash(byte[] hash);
+
+    void saveBlock(BLK block, BigInteger cummDifficulty, boolean mainChain);
+
+    BigInteger getTotalDifficulty();
 
 }
