+++ modAionImpl/src/org/aion/zero/impl/db/AionBlockStore.java
@@ -27,7 +27,7 @@ import org.aion.base.util.ByteUtil;
 import org.aion.base.util.Hex;
 import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
-import org.aion.mcf.db.AbstractPowBlockstore;
+import org.aion.zero.db.AbstractPowBlockstore;
 import org.aion.mcf.ds.DataSourceArray;
 import org.aion.mcf.ds.ObjectDataSource;
 import org.aion.mcf.ds.Serializer;
