+++ modMcf/src/org/aion/mcf/evt/EvtCb.java
@@ -25,6 +25,6 @@
 package org.aion.mcf.evt;
 
 @FunctionalInterface
-public interface EvtCb<T extends EvtData> {
+public interface EvtCb<T> {
     void call(T t);
 }
